back-photo-search/requirements.txt
Flask==3.1.0
marshmallow==3.23.1
python-dotenv==1.0.1
requests-cache==1.2.1
Flask-CORS==5.0.0
